Add XForwardedFiltered class
There are some circumstances in which the number of internal proxies
in front of a service is not fixed.  An example is an auth_request
handler for NGINX behind a Kubernetes NGINX ingress controller.  When
called as an auth_request handler, there are two layers of proxying.
When called directly, there is only one layer of proxying.  In this
situation, there is no good choice of num for XForwardedRelaxed.

XForwardedFiltered adds a trusted proxy list similar to
XForwardedStrict, but is not strict about the exact number of proxies
in front of the service.  Instead, it walks X-Forwarded-For from the
right side and drops entries that match the trusted list.  It then
sets the request attributes based on the first non-trusted IP that was
found.  If all IPs are in the trusted list, it uses the left-most
entry.

The current Kubernetes NGINX ingress controller only calculates a full
X-Forwarded-For header and provides only a single value in
X-Forwarded-Proto, so this implementation silently copes with that and
returns that single value if there are not enough entries in
X-Forwarded-Proto to find the entry corresponding to X-Forwarded-For.

Closes #153.